refactor: Entity에 toString 메서드 추가
