+++ src/main/java/org/example/repository/WiseSayingRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 import org.example.config.ConfigReader;
-import org.example.dto.WiseSaying;
+import org.example.domain.WiseSaying;
 import org.example.config.FileToWiseSaying;
 
 public class WiseSayingRepository {
